Modify the size of labels to fix warnings on GPU
When I was training the model on CPU, everything's OK. When I tried to train on GPU, it incessantly print the following warning message.
`[W ..\..\aten\src\ATen\native\Resize.cpp:19] Warning: An output with one or more elements was resized since it had shape [64, 1, 1, 1], which does not match the required output shape [64, 1, 1, 64].This behavior is deprecated, and in a future PyTorch release outputs will not be resized unless they have zero elements. You can explicitly reuse an out tensor t by resizing it, inplace, to zero elements with t.resize_(0). (function resize_output)`
The warning happens each time backward() function is called. It seems that tensors of size {64} will not be compatible in calculation with tensors of size {64,1,1,1} in the future PyTorch release. After I modified the size of fake_labels and real_labels to {64,1,1,1}, the warning messages disappeared.